smtp - replace submodule with base64c library
This is simple library intended for just base64 encoding and decoding.
It is in no way beter than replaced one with exception that it is
packaged as library and not just as repository to be included. It makes
this project cleaner but of course adds external dependency that user
has to install.