build: reogranize repository
This moves source files to dedicated directory.

This also removes gen-gperf target as its usage is questionable. It is
always possible to build selected C file and usage to generate all gperf
C files is pretty much not there. In the full build the files are pulled
in as needed.